backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-cors==0.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
alembic==1.12.1

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & AI
langchain==0.1.0
langgraph==0.0.1
langchain-openai==0.0.7
langchain-groq==0.0.1
openai==1.3.8
groq==0.4.2
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
azure-cognitiveservices-speech==1.31.0

# Voice & Audio
pyttsx3==2.90
pyaudio==0.2.13
wave==0.0.2

# Emergency & Location
twilio==8.10.0
google-maps-services==1.10.0

# Utilities
requests==2.31.0
httpx==0.25.2
redis==5.0.1
aioredis==2.0.1

# Logging & Monitoring
python-logging-loki==0.3.2
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# WSGI/ASGI
gunicorn==21.2.0
